fix done

+++ GameSDLCppDemo/CakeObject.h
@@ -17,8 +17,8 @@ public:
     void Reset(const int& yboder);
     bool CakeExit(const int& ye);
 private:
-    int x_val_; // This member variable is not used in the provided implementation
-    int y_val_; // This member variable is used for vertical movement
+    double x_val_; // This member variable is not used in the provided implementation
+    double y_val_; // This member variable is used for vertical movement
 };
 
 #endif
